Quadratic Equation
Comment of how the function works has been described for both the negative and positive values of the quadratic equations. in the arithmetic operators.

The cases 7 and 8 in the Calculator file have 3 numbers that have been added relative to the relevant equation. These values are for both the positive and negative parts of the quadratic equation.